pyproject.toml
[tool.poetry]
name = "RAG-llm"
version = "0.1.0"
description = ""
authors = ["sarohy <a.h.sarohy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.1"
pymongo = "^4.8.0"
python-dotenv = "^1.0.1"
PyPDF2 = "^3.0.1"
beautifulsoup4 = "^4.12.3"
fpdf = "^1.7.2"
jose = "^1.0.0"
requests = "^2.32.3"
pycryptodome = "^3.20.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
uvicorn = "^0.30.6"
python-multipart = "^0.0.9"
motor = "^3.5.1"
scikit-learn = "^1.5.1"
transformers = "^4.44.1"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
